requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
pydantic==2.10.3
pydantic-settings==2.7.0
torch==2.11.0   
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4
tiktoken==0.8.0
httpx==0.28.1
python-dotenv==1.0.1
python-multipart
pypdf
python-docx
openpyxl
tavily-python
